Doctors Full

+++ hospital/src/main/java/com/example/hospital/repository/IDoctorRepository.java
@@ -1,7 +1,8 @@
 package com.example.hospital.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.hospital.model.Doctor;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDoctorRepository extends JpaRepository<Doctor, Integer> {
+public interface IDoctorRepository extends JpaRepository<Doctor, Long> {
+    // Puedes agregar métodos personalizados aquí si es necesario
 }
